Fix cedilla and plusmn translations
The cedilla in LaTeX is actually `\c{c}` (see, e.g.
https://en.wikibooks.org/wiki/LaTeX/Special_Characters#Escaped_codes) and
the `\pm` command needs to be wrapped in dollar signs if it's used outside
of math mode (i.e. outside an equation, $$, or similar environment).  This
commit fixes the issue so that if the test file (`t/test_file.pod`) is run
through `ppod2latex` then it produces LaTeX which can be processed properly
by `latex` or `pdflatex`.